requirements.txt
discord.py
flask
python-dotenv
Pillow
requests
waitress
